Trait blockz::configuration::Configuration

pub trait Configuration {
    type Opts: Send;
    type Result: Send;
    fn load<'async_trait>(
        opts: Self::Opts
    ) -> Pin<Box<dyn Future<Output = Self::Result> + Send + 'async_trait>>
    where
        Self: 'async_trait;
}

This is supported on crate feature configuration only.

Common behaviour of configurations.

Associated Types

type Opts: Send

The type of options container this Configuration accepts for the purpose of loading the configuration.

type Result: Send

The result type that can be produced by loading the configuration.

Required methods

fn load<'async_trait>(
opts: Self::Opts
) -> Pin<Box<dyn Future<Output = Self::Result> + Send + 'async_trait>> where
Self: 'async_trait,

Load the configuration.